[#16] Feat: 가격 입력/확인 페이지 render
