+++ src/pages/PriceCheck.tsx
@@ -1,0 +1,7 @@
+export default function PriceCheck() {
+    return (
+        <div>
+            <h1>PriceCheck</h1>
+        </div>
+    )
+}
